rename: hook file 이름 변경
